+++ CHANGES.md
@@ -3,6 +3,7 @@
 ## 10th July 2024 - 0.11.8
 
 * Boilerplate files improvements
+* Moved bundled **BDUT** header(s) into separate directory
 
 
 ## 0.11.7 - 12th May 2024